+++ bin/static/security.js
@@ -5186,4 +5186,4 @@
   };
 })();
 /* ui-bundle: 20a79c744bffe82c21b79d103546d78595f23f9f117db57365c009a88216cfdd */
-/* ui-sources: ea07ded1b0d1d804c23b09e35ca69c05ecfbe4496d78c9d6bff1953a08686302 */
+/* ui-sources: 134b851fbcb498cab658c010ab60be13d9f3103555f575a9d140a3c3eccfbfdb */